Finally found the link between gRPC and ProtocolBuffers
I finally found the link and stack trace between gRPC and ProtocolBuffers. There is just a "weak" connection between BlockingUnaryCallImpl object and the CallOpSendMessage object because the former does use a CallOpSendMessage within a CallOpSet; however, the file is not included. Perhaps, somewhere in the directory, the CallOps are forward declared. Regardless, we will use this information for instrumentation and perhaps it can help with creating the configuration file.